% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD_correlator.R
\name{AD_correlator}
\alias{AD_correlator}
\title{AnnotateLater}
\usage{
AD_correlator(
  CountDF,
  dAD_fitres,
  SE,
  CustomCor = NULL,
  MinCount = 0,
  method = "spearman"
)
}
\description{
AnnotateLater
}
