% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdvancedPvalcomb.R
\name{AdvancedPvalcomb}
\alias{AdvancedPvalcomb}
\title{Combines SNP-level dAD results to gene-level.}
\usage{
AdvancedPvalcomb(DATAS, SE, nperm = 10000)
}
\arguments{
\item{DATAS}{List. A list containing four elements:
A vector of (unique) gene names whose SNP-level results are being combined (1).
A list of subsets of the `dAD_analysis` output dataframe containing, for every gene (see first list-element), the SNP-level results for SNPs assigned to said gene (2).
A list of lists of dataframes containing, for every gene, a list of per-SNP dataframes containing data for control samples
(columns "Sample", "Outlier", "ref_count", and "var_count"; respectively sample-IDs, an Outlier indicator (0 if no outlier, 1 if so), reference-, and variant allele counts) (3).
A list of lists of dataframes containing, for every gene, a list of per-SNP dataframes containing data for case samples
(columns "Sample", "Outlier", "ref_count", and "var_count"; respectively sample-IDs, an Outlier indicator (0 if no outlier, 1 if so), reference-, and variant allele counts) (4).}

\item{SE}{Number. Sequencing error.}

\item{nperm}{Integer. Number of iterations used for correlation-corrected Lancaster p-value combination, which estimates the dependence between samples through a (large) number of sample permutations (default 10000).}
}
\value{
A list containing the following components:
\item{FisherPVAL_vec}{Vector containing, for all genes, combined p-values using the Fisher combination method.}
\item{LancasterPVAL_vec}{Vector containing, for all genes, combined p-values using the Lancaster combination method.}
\item{CorLancasterPVAL_vec}{Vector containing, for all genes, combined p-values using the dependence(correlation)-corrected Lancaster combination method.}
\item{CauchyPVAL_vec}{Vector containing, for all genes, combined p-values using the Cauchy combination method.}
\item{CLPV_OffDiagSum_vec}{Vector containing, for all genes, the sum of between-sample correlations as estimated during the dependence corrected Lancaster combination method;
a sort of measure for the amount of between-sample dependence present amongst SNPs.}
}
\description{
\code{AdvancedPvalcomb} relies on the output of maelstRom's `dAD_analysis` function, applied on SNP-level data, and combines them to SNP-level p-values.
Written to be run in parallel through the parallel's package `parLapply` function.
}
