% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllelicMeta_est.R
\name{AllelicMeta_est}
\alias{AllelicMeta_est}
\title{Estimate sequencing error rate and inbreeding using an Expectation-Maximisation algorithm}
\usage{
AllelicMeta_est(
  ref_counts,
  var_counts,
  deltaF = 10^-8,
  maxIT = 100,
  SE_prior = 0.002,
  F_inbr_prior = NULL,
  HetProb = 0.5
)
}
\arguments{
\item{ref_counts}{Numeric vector. Reference counts.}

\item{var_counts}{Numeric vector. Variant counts.}

\item{deltaF}{Number. Expectation-Maximisation threshold, minimal difference between
two consecutive iterations (default is 1e-08).}

\item{maxIT}{Number. Maximum number of iterations of the Expectation-Maximisation algorithm (default is 100).}

\item{SE_prior}{Number. Initial estimate of the sequencing error rate (default is 0.002).}

\item{F_inbr_prior}{Number. Initial estimate of the inbreeding coefficient used for calculating initial genotype frequencies (default is NULL, in which case the initial genotype frequences all get set to 1/3).}

\item{HetProb}{Number. Allelic bias in heterozygotes (expected reference over total allele count in RNAseq data; default is 0.5}
}
\value{
A list containing the following components:
\item{allelefreq}{The estimated allele frequency.}
\item{SE}{The estimated sequencing error rate.}
\item{F_inbr}{The estimated inbreeding coefficient.}
\item{genotypes}{The most likely genotype (rr, rv, vv) of each sample.}
\item{genoprobs}{The genotype probabilties (p(rr), p(rv), p(vv)) for each sample.}
\item{nrep}{The number of iterations}
}
\description{
\code{AllelicMeta_est} calculates sequencing error rates, inbreeding
    coefficients, allele frequencies, genotypes and genotype probabilities from \code{ref_counts} and \code{var_counts}
    for a specific locus. To this end it fits a (regular) binomial mixture model with no allelic bias in heterozygotes (both alleles equally likely).
    This somewhat simplistic model (no overdispersion parameter, no allelic bias) makes for an inferior but very fast fit,
    which is ideal to get rough estimates of the allelic population metaparameters (sequencing error rate and inbreeding coefficient) by e.g. taking their mean or median across a large amount of loci.
    The per-sample allele frequencies and genotype (probabilities) are less reliable and, as such, not this function's main use.
}
\examples{
AllelicMeta_est(c(5, 8, 10, 3, 5, 6, 23), c(8, 8, 6, 4, 4, 10, 0))
AllelicMeta_est(c(5, 0, 0, 3, 5, 1, 23), c(1, 8, 6, 2, 0, 10, 0),
    SE_prior = 0.2, F_inbr_prior = 0.1)
}
