% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{allpass_filter}
\alias{allpass_filter}
\title{Schroeder allpass filter}
\usage{
allpass_filter(x, delay, g)
}
\arguments{
\item{x}{Input signal.}

\item{delay}{Delay in samples.}

\item{g}{Gain.}
}
\value{
The filtered signal.
}
\description{
Allpass used by the Schroeder reverb: a convolution feedforward stage into a recursive stage.
}
\examples{
allpass_filter(c(1, rep(0, 99)), 10, 0.5)
}
