% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini.R
\name{atom_pattern}
\alias{atom_pattern}
\title{Pattern holding one value over a span}
\usage{
atom_pattern(value, span)
}
\arguments{
\item{value}{A hap value list.}

\item{span}{A TimeSpan.}
}
\value{
A pattern.
}
\description{
A single hap with the given value sounding over `span`. The span is snapshotted so loop-built closures do not capture a changing variable.
}
\examples{
query_arc(atom_pattern(list(sound = "bd"), timespan(0, 1)), 0, 1)
}
