% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators.R
\name{attack}
\alias{attack}
\title{Set the attack time}
\usage{
attack(pat, x)
}
\arguments{
\item{pat}{A pattern.}

\item{x}{Attack in seconds.}
}
\value{
A pattern.
}
\description{
Attack time in seconds for synth voices.
}
\examples{
note("c4") |> attack(0.05)
}
