% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{audio_tag}
\alias{audio_tag}
\title{Inline audio HTML tag}
\usage{
audio_tag(samples, sr = SR)
}
\arguments{
\item{samples}{Mono vector or 2 x N stereo matrix.}

\item{sr}{Sample rate, default 44100.}
}
\value{
An HTML `<audio>` string.
}
\description{
Render samples to a WAV in memory and wrap it in an `<audio>` tag with a data URI.
}
\examples{
audio_tag(c(0.1, -0.2, 0.3))
}
