% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaBinomGenotyping.R
\name{BetaBinomGenotyping}
\alias{BetaBinomGenotyping}
\title{multi-locus wrapper function of EMfit_betabinom_robust for the purpose of genotyping}
\usage{
BetaBinomGenotyping(
  DataList,
  allelefreq = 0.5,
  SE,
  inbr = 0,
  dltaco = 10^-6,
  HWE = FALSE,
  p_InitEst = FALSE,
  ThetaInits = "moment",
  ReEstThetas = "moment",
  NoSplitHom = TRUE,
  NoSplitHet = TRUE,
  ResetThetaMin = 10^-10,
  ResetThetaMax = 10^-1,
  DistRob = "Cook",
  CookMargin = 5,
  LikEmpNum = 1000,
  LikMargin = 0,
  NumHetMin = 5,
  MaxOutFrac = 0.5,
  thetaTRY = c(10^-1, 10^-3, 10^-7),
  fitH0 = TRUE
)
}
\arguments{
\item{DataList}{List of dataframes. Each dataframes should at least contain the columns "ref", ref_count", "var", "var_count" and "est_SE",
which should be the case if following the vignette up to this function's appearance.}

\item{allelefreq, SE, inbr, dltaco, HWE, p_InitEst, ThetaInits, ReEstThetas, NoSplitHom, NoSplitHet, ResetThetaMin, ResetThetaMax, DistRob, CookMargin, LikEmpNum, LikMargin, NumHetMin, MaxOutFrac, thetaTRY, fitH0}{All remaining parameters of \code{EMfit_betabinom_robust}}
}
\value{
A list containing the following components:
\item{DataList_out}{The updated DataList, corresponding to the \code{data_hash} output of \code{EMfit_betabinom_robust}}
\item{Geno_AB_res}{A dataframe containing per-locus results of \code{EMfit_betabinom_robust} EM-fit as well as some other metrics, namely:}
   \itemize{
     \item{position}{The locus' name, according to \code{names(DataList)}}
     \item{probshift}{Fitted reference allele fraction in RNAseq reads, indicating allelic bias when different from 0.5}
     \item{LRT}{The likelihood ratio test statistic, testing for significant allelic bias}
     \item{p}{The likelihood ratio test p-value, testing for significant allelic bias}
     \item{quality}{Equals "!" if the sample contains no fitted heterozygotes, otherwise ""}
     \item{allele_frequency}{estimated reference allele frequency in the population}
     \item{reference}{reference allele nucleotide}
     \item{variant}{variant allele nucleotide}
     \item{est_SE}{per-locus sequencing error rate estimate, as outputted earlier by \code{AllelicMeta_est}}
     \item{coverage}{median coverage across samples}
     \item{nr_samples}{number of samples covering this locus with at least one reference- or variant-count}
     \item{median_AB}{median allelic bias, as outputted by \code{median_AB}}
     \item{rho_rr}{fitted reference homozygous fraction in the population}
     \item{rho_rv}{fitted heterozygous fraction in the population}
     \item{rho_vv}{fitted variant homozygous fraction in the population}
     \item{theta_hom}{fitted overdispersion parameter for the homozygous PMFs}
     \item{theta_het}{fitted overdispersion parameter for the heterozygous PMF}
     \item{theta_hom_NoShift}{fitted overdispersion parameter for the homozygous PMFs assuming no allelic bias}
     \item{theta_het_NoShift}{fitted overdispersion parameter for the heterozygous PMF assuming no allelic bias}
     \item{Chi2PVAL}{p-value of a chi square test assessing Hardy-Weinberg-Equilibrium on the locus given the inbreeding coefficient metaparameter; see \code{HWE_chisquared}}
     \item{Chi2STAT}{test statistic of a chi square test assessing Hardy-Weinberg-Equilibrium on the locus given the inbreeding coefficient metaparameter; see \code{HWE_chisquared}}
   }
}
\description{
\code{BetaBinomGenotyping} is a wrapper function of \code{EMfit_betabinom_robust} allowing multiple loci to be given as input (as lists).
Besides calling the latter to perform a EM-fit and update the inputted per-locus dataframes as listed in \code{EMfit_betabinom_robust} output,
it also returns a results-dataframe with one row for every locus, containing genotyping- and Allelic Bias-detection results.
This function is mainly useful when following MAGE's vignette step-by-step, as its input requirement are rather stictly dependent on performing all previous steps in that pipeline.
See the help page of \code{EMfit_betabinom_robust} for more information on these outputs
}
