% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaBinomMix_LLDeriv_transfo.R
\name{BetaBinomMix_LLDeriv_transfo}
\alias{BetaBinomMix_LLDeriv_transfo}
\title{Beta-binomial mixture model log-likelihood derivatives to its parameters involved in optimization}
\usage{
BetaBinomMix_LLDeriv_transfo(
  ref_counts,
  var_counts,
  probshift,
  SE,
  prr,
  pvv,
  prv,
  theta_hom,
  theta_het,
  Der1,
  Der2 = NULL
)
}
\arguments{
\item{ref_counts}{Number or Numeric vector Reference count(s).}

\item{var_counts}{Number or Numeric vector. Variant count(s).}

\item{probshift}{Number. The reference allele fraction in heterozygotes, indicating allelic bias when deviating from 0.5}

\item{SE}{Number. Sequencing error rate.}

\item{prr}{Number. Reference homozygote genotype probability of the locus.}

\item{pvv}{Number. Variant homozygote genotype probability of the locus.}

\item{prv}{Number. Heterozygote genotype probability of the locus.}

\item{theta_hom}{Number. The dispersion parameter of the homozygous peaks.}

\item{Der1}{String. Derivate with respect to this parameter.}

\item{Der2}{String. Derivate with respect to this parameter as well; optional.}

\item{isCase}{Number or Numeric vector. Equals 1 for cases and 0 for controls.}
}
\value{
A log-likelihood derivative
}
\description{
\code{BetaBinomMix_LLDeriv} Lorem ipsum
}
