% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth2.R
\name{bitcrush}
\alias{bitcrush}
\title{Bitcrush}
\usage{
bitcrush(x, bits)
}
\arguments{
\item{x}{Input signal.}

\item{bits}{Bit depth in [1, 16].}
}
\value{
The quantized signal.
}
\description{
Quantize amplitudes to `bits` bits.
}
\examples{
bitcrush(c(-1, -0.3, 0.3, 1), 4)
}
