% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{bjorklund}
\alias{bjorklund}
\title{Bjorklund Euclidean rhythm}
\usage{
bjorklund(hits, steps)
}
\arguments{
\item{hits}{Number of pulses.}

\item{steps}{Number of slots.}
}
\value{
A logical vector of length `steps`.
}
\description{
Distribute `hits` pulses as evenly as possible over `steps` slots.
}
\examples{
bjorklund(3, 8)
}
