% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_p_gene.R
\name{combine_p_gene}
\alias{combine_p_gene}
\title{Combine p-values of SNPs per gene.}
\usage{
combine_p_gene(p_values, weights = NULL, method = "geometric", r = NULL)
}
\arguments{
\item{p_values}{Numeric vector. P-values to combine.}

\item{weights}{Numeric vector. Weights given to each observation (corresponds to how much an observation occurs, but can be non-integer as well).}

\item{method}{String. Specifies the method for combining p-values, the options being "arithmetic" (mean), "geometric" (mean),
"harmonic" (mean), "minimum" (p-value), "maximum" (p-value) or "GMP" (Generalized Mean P-value).
In case of the latter, a value for r must be specified, and the GMP is then calculated as (sum(weights * p_values^r)/sum(weights))^(1/r), see \insertCite{wilson2020generalized}{MAGE}.}

\item{r}{Numeric. r-value to be used in the calculation of the GMP, should that option be chosen as the method.}
}
\value{
Combined p-value of a gene.
}
\description{
\code{combine_p_gene} combines per-SNP p-values into per-gene p-values;
the method for doing so can be chosen via input
}
\examples{
combine_p_gene(c(0.015, 0.414, 8.47E-03, 7.43E-03, 0.574, 0.837))
combine_p_gene(c(0.063, 0.725, 0.657, 0.378, 0.291))
}
\references{
\insertAllCited{}
}
