% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CppHom_Optim}
\alias{CppHom_Optim}
\title{DO SOMETHING}
\usage{
CppHom_Optim(
  ThetaHomStart,
  SE,
  ref_counts,
  var_counts,
  spr,
  spv,
  MemLim = 2048L,
  Xtra = 7L,
  step_size = 0.01,
  tol = 0.1,
  epsabs = 0.001
)
}
\description{
DO SOMETHING
}
