% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dAD_analysis.R
\name{dAD_analysis}
\alias{dAD_analysis}
\title{Performs a differential Allelic Divergence analysis on a control- and case-dataset}
\usage{
dAD_analysis(
  DataList,
  SE,
  inbr = 0,
  allelefreq = 0.5,
  dltaco = 10^-6,
  HWE = FALSE,
  p_InitEst = FALSE,
  ThetaInits = "moment",
  ReEstThetas = "moment",
  NoSplitHom = TRUE,
  NoSplitHet = TRUE,
  ResetThetaMin = 10^-10,
  ResetThetaMax = 10^-1,
  DistRob = "Cook",
  CookMargin = 5,
  LikEmpNum = 1000,
  LikMargin = 0,
  NumHetMin = 5,
  MaxOutFrac = 0.5,
  thetaTRY = c(10^-1, 10^-3, 10^-7)
)
}
\arguments{
\item{SE}{Number. Sequencing error rate population metaparameter.}

\item{inbr}{Number. Inbreeding coefficient population metaparameter.}

\item{allelefreq, dltaco, HWE, p_InitEst, ThetaInits, ReEstThetas, NoSplitHom, NoSplitHet, ResetThetaMin, ResetThetaMax, DistRob, CookMargin, LikEmpNum, LikMargin, NumHetMin, MaxOutFrac, thetaTRY}{Remaining
control parameters to be passed to \code{EMfit_betabinom_robust}, \code{EMfit_betabinom}, and \code{EMfit_betabinom_popcomb}. See their respective documentation
for more information. Usually, these parameter's default values are fine.}

\item{Datalist}{A list containing two lists of dataframes, which in turn contain control- and case-data (in that order). Subsequent entries in both lists
correspond to subsequent loci (lists must be the same size and named per-locus). Each of the dataframes should at least contain both a \code{ref_count} and
\code{var_count} column.}
}
\value{
A dataframe with one row per locus, containing for that locus in its columns:
\item{LocName}{Locus name.}
\item{PiFitH0}{Heterozygous pi parameter (reference allele fraction) for the joint fit sharing all parameters.}
\item{PiFitH1}{Heterozygous pi parameter (reference allele fraction) for the fit allowing differential Allelic Divergence.}
\item{ThetaHetH0}{Heterozygous theta parameter (overdispersion) for the joint fit sharing all parameters.}
\item{ThetaHetCTRL}{Heterozygous theta parameter (overdispersion) of controls for the fit allowing differential Allelic Divergence.}
\item{ThetaHetCASE}{Heterozygous theta parameter (overdispersion) of cases for the fit allowing differential Allelic Divergence.}
\item{RhoHetH0}{Heterozygous rho parameter (overdispersion, ranging from 0 to 1) for the joint fit sharing all parameters.}
\item{RhoHetCTRL}{Heterozygous rho parameter (overdispersion, ranging from 0 to 1) of controls for the fit allowing differential Allelic Divergence.}
\item{RhoHetCASE}{Heterozygous rho parameter (overdispersion, ranging from 0 to 1) of cases for the fit allowing differential Allelic Divergence.}
\item{NumHetCTRL}{Estimate number of heterozygotes in controls, according to its separate outlier-detection fit from step 1 above. Outliers are themselves genotyped and counted towards this number, just not used in the actual fitting procedure.}
\item{NumHetCASE}{Estimate number of heterozygotes in cases, according to its separate outlier-detection fit from step 1 above. Outliers are themselves genotyped and counted towards this number, just not used in the actual fitting procedure.}
\item{RobFlagCTRL}{The \code{RobFlag} output of the separate outlier-detection fit from step 1 on controls, see \code{EMfit_betabinom_robust}'s documentation.}
\item{RobFlagCASE}{The \code{RobFlag} output of the separate outlier-detection fit from step 1 on cases, see \code{EMfit_betabinom_robust}'s documentation.}
\item{HWECTRL}{HWE chi squared p-value for controls, see \code{HWE_chisquared}. Performed on genotyping data according to the outlier-detection fit of step 1.}
\item{HWECASE}{HWE chi squared p-value for cases, see \code{HWE_chisquared}. Performed on genotyping data according to the outlier-detection fit of step 1.}
\item{CovCTRL_mean}{Mean coverage (reference + variant count) across control data.}
\item{CovCASE_mean}{Mean coverage (reference + variant count) across case data.}
\item{CovCTRL_med}{Median coverage (reference + variant count) across control data.}
\item{CovCASE_med}{Median coverage (reference + variant count) across case data.}
\item{NumOutCTRL}{Number of outliers detected in control data.}
\item{NumOutCASE}{Number of outliers detected in case data.}
\item{QualityCTRL}{Quality flag returned by \code{EMfit_betabinom_robust} (see documentation) of the outlier-detection fit on control data (step 1 above).}
\item{QualityCASE}{Quality flag returned by \code{EMfit_betabinom_robust} (see documentation) of the outlier-detection fit on case data (step 1 above).}
\item{pr}{Reference homozygote fraction according to the differential Allelic Divergence accomodating fit (different heterozygous overdispersion parameters).}
\item{prv}{Heterozygote fraction according to the differential Allelic Divergence accomodating fit (different heterozygous overdispersion parameters).}
\item{pv}{Variant homozygote fraction according to the differential Allelic Divergence accomodating fit (different heterozygous overdispersion parameters).}
\item{ThetaHomH0}{Homozygous theta parameter (overdispersion) for the joint fit sharing all parameters.}
\item{ThetaHomH1}{Homozygous theta parameter (overdispersion) for the fit allowing differential Allelic Divergence.}
}
\description{
\code{dAD_analysis} is a wrapper function bundling \code{EMfit_betabinom_robust}, \code{EMfit_betabinom}, \code{EMfit_betabinom_popcomb},
\code{pmf_betabinomMix}, and \code{HWE_chisquared} to run the entire differential Allelic Divergence detection analysis as discussed in the package vignette.
To this end, it goes over the following steps:
\enumerate{
  \item Calling \code{EMfit_betabinom_robust} on both the control- and case-dataset separately for the sole purpose of outlier detection;
  detected outliers in both datasets play no part in future Expectation-Maximization-fits.
  \item Perform one fit on the joint (case+control) data using \code{EMfit_betabinom}, i.e. assuming all population parameters are shared including
  the heterozygous overdispersion parameter, i.e. assuming no differential Allelic Divergence.
  \item Perform a fit on the control- and case-data using \code{EMfit_betabinom_popcomb}, which assumes all population parameters are shared
  except for the heteroyzous overdispersion parameter, i.e. a fit accomodating differential Allelic Divergence
  \item Statistically test for differential Allelic Divergence using a likelihood ratio test with one degree of freedom comparing the
  fits from the two previous points.
  \item Filling out the results dataframe \code{dAD_res}.
}
}
