% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dBetaBin_cppi}
\alias{dBetaBin_cppi}
\title{dBetaBinom implementation for internal use by C++}
\usage{
dBetaBin_cppi(M, N, PI, THETA, LOG, MemLim, Xtra)
}
\description{
dBetaBinom implementation for internal use by C++
}
