% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dBetaBinom_cpp_old}
\alias{dBetaBinom_cpp_old}
\title{Exact beta-binomial density using sums.}
\usage{
dBetaBinom_cpp_old(ms, ns, pi, theta, LOG)
}
\arguments{
\item{ms}{Numeric vector. Vector of number of successes}

\item{ns}{Numeric vector. Vector of number of trials}

\item{pi}{Number. Probability of success; \code{0 >= pi >= 1}}

\item{theta}{Number. Dispersion parameter; \code{0 >= theta > +Inf}}

\item{LOG}{Logical. if TRUE, return log-densities}
}
\value{
A numeric vector of the same length as ms and ns, containing (log-)beta-binomial densities
}
\description{
\code{dBetaBinom_cpp_old} calculates the beta-binomial density via
a number of sums, which is slow for high-value data
but fast for low-value data.
This function is called by \code{dBetaBinom}
if necessary, and should not be called outside of this.
}
