% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dBetaBinom_MP}
\alias{dBetaBinom_MP}
\title{Exact beta-binomial density using a multiprecision library.}
\usage{
dBetaBinom_MP(ms, ns, piX, thetaX, LOG, NecPres)
}
\arguments{
\item{ms}{Numeric vector. Vector of number of successes}

\item{ns}{Numeric vector. Vector of number of trials}

\item{piX}{Number. Probability of success; \code{0 >= piX >= 1}}

\item{thetaX}{Number. Dispersion parameter; \code{0 >= thetaX > +Inf}}

\item{LOG}{Logical. if TRUE, return log-densities}

\item{NecPres}{Number. Necessary Precision, i.e. number of bits, for an accurate density calculatation, as determined by the function \code{dBetaBinom}}
}
\value{
A numeric vector of the same length as ms and ns, containing (log-)beta-binomial densities
}
\description{
\code{dBetaBinom_MP} calculates the beta-binomial density while
avoiding numerical mistakes (catastrophic cancellations) due to 
extreme parameter values. This function is called by \code{dBetaBinom}
if necessary, and should not be called outside of this.
}
