% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators.R
\name{degrade_by}
\alias{degrade_by}
\title{Randomly drop events by an amount}
\usage{
degrade_by(pat, amount)
}
\arguments{
\item{pat}{A pattern.}

\item{amount}{Drop probability in [0, 1].}
}
\value{
A pattern.
}
\description{
Alias of `degrade` with the amount required.
}
\examples{
query_arc(degrade_by(s("bd*8"), 0.25), 0, 1)
}
