% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/dpqrBetaBinom.R
\name{dBetaBinom}
\alias{dBetaBinom}
\alias{dpqrBetaBinom}
\alias{pBetaBinom}
\alias{qBetaBinom}
\alias{rBetaBinom}
\title{The Beta-binomial distribution}
\usage{
dBetaBinom(ms, ns, pi, theta, LOG = FALSE, MemLim = 2048L, Xtra = 7L)

pBetaBinom(ms, ns, pi, theta, lower.tail = TRUE, log.p = FALSE)

qBetaBinom(p, ns, pi, theta, lower.tail = TRUE, log.p = FALSE)

rBetaBinom(ns, pi, theta)
}
\arguments{
\item{ms}{Numeric vector. Vector of number of successes}

\item{ns}{Numeric vector. Vector of number of trials}

\item{pi}{Number. Probability of success; \code{0 >= pi >= 1}}

\item{theta}{Number. Dispersion parameter; \code{0 >= theta > +Inf}}

\item{LOG}{Logical. If TRUE, \code{dBetaBinom} returns log-densities.}

\item{MemLim}{Integer. The memory limit, in bits, of numbers used during the calculation of the density.
For extreme data/parameter values, it may be necesssary to go beyond R's \code{double} memory, i.e. 53 bits,
in order to get correct results. However, this argument prevents boundless memory usage.
In case this limit is reached, \code{dBetaBinom} falls back on the regular \code{dbinom} function,
as this should only happen for extremely close-to-zero theta-values anyway.
A warning message is generated if this occurs.}

\item{Xtra}{Integer. An internal control parameter, determines the number of decimal places that are to be
stored correctly in memory during calculation of beta-binomial densities, upon which the number of
bits to be used per number depends. It's not recommended to change this from the default.}

\item{log.p}{Logical. If TRUE, probabilities p returned and accepted by \code{pBetaBinom} and \code{qBetaBinom} respectively are given as log(p).}
}
\value{
\code{dBetaBinom} gives the density, \code{pBetaBinom} gives the distribution function, \code{qBetaBinom} gives the quantile function and \code{rBetaBinom} generates random deviates. Where applicable, output has the same length as \code{ms} and/or \code{ns}.
}
\description{
Density, distribution function, quantile function and random generation for the beta-binomial distribution with parameters \code{pi} and \code{theta}.
These can be interpreted as an expected probability of success and a dispersion parameter respectively,
and can respectively be expressed in terms of the common \code{alpha, beta} parameterization as \code{pi = alpha/(alpha+beta)} and \code{theta = 1/(alpha+beta)}.
}
