% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMfit_betabinom_SEslack.R
\name{EMfit_betabinom_SEslack}
\alias{EMfit_betabinom_SEslack}
\title{Models allelic RNAseq counts using an expectation-maximization fit of a beta-binomial mixture distribution, allowing slight deviations in the Sequencing Error (SE) metaparameter.}
\usage{
EMfit_betabinom_SEslack(
  data_counts,
  allelefreq = 0.5,
  SE,
  inbr = 0,
  dltaco = 10^-6,
  HWE = FALSE,
  p_InitEst = FALSE,
  ThetaInits = "moment",
  ReEstThetas = "moment",
  NoSplitHom = TRUE,
  NoSplitHet = TRUE,
  ResetThetaMin = 10^-10,
  ResetThetaMax = 10^-1,
  thetaTRY = c(10^-1, 10^-3, 10^-7),
  fitH0 = TRUE,
  SESlack = 10,
  probshift_InitEst = TRUE
)
}
\description{
\code{EMfit_betabinom_SEslack} is identical to `EMfit_betabinom`, but allows for the Sequencing Error (SE) parameter, used in the pi-parameter of homozygous beta-binomial peaks,
to deviate from its given value slightly, up to a maximum of "SESlack" (input) times SE. For an explanation of the remaining inputs and the outputs, see `EMfit_betabinom`.
THIS FUNCTION IS CURRENTLY NOT USED AS ITS PERFORMANCE ON ACTUAL DATA PROVED UNSATISFACTORY.
}
