% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMfit_binom.R
\name{EMfit_binom}
\alias{EMfit_binom}
\title{Models allelic RNAseq counts using an expectation-maximization fit of a binomial mixture distribution}
\usage{
EMfit_binom(
  data_counts,
  SE,
  allelefreq = 0.5,
  inbr = 0,
  dltaco = 0.001,
  HWE = FALSE,
  p_InitEst = FALSE
)
}
\arguments{
\item{data_counts}{Data frame. Data frame of a SNP with reference and variant counts ("ref_count" and
"var_count", respectively) for each sample ("sample").}

\item{SE}{Number. Sequencing error rate.}

\item{allelefreq}{Number. Allele frequency. Only used when pInitEst is TRUE (default = 0.5)}

\item{inbr}{Number. Degree of inbreeding (default = 0).}

\item{dltaco}{Number. Minimal difference between 2 iterations (default = 0.001).}

\item{HWE}{Logical. Should HWE be used for allele frequency estimation, not recommended (default = FALSE).}

\item{p_InitEst}{Logical. Calculate initial estimates of pr, pv and prv from allelefreq, not recommended (default = FALSE).}
}
\value{
A list containing the following components:
\item{AB}{The estimated allelic shift.}
\item{AB_lrt}{The test statistic of the likelihood ratio test.}
\item{AB_p}{The p-value of the likelihood ratio test.}
\item{GOF}{The goodness-of-fit value based on the corrected likelihood.}
\item{nrep}{The number of iterations.}
\item{quality}{Indicates the quality of a locus. An "!" indicates bad data or a bad fit due to no apparent heterozygosity.}
\item{rho_vv}{The variant (homozygote) genotype probability for the SNP.}
\item{rho_rr}{The reference (homozygote) genotype probability for the SNP.}
\item{rho_rv}{The heterozyous genotype probability for the SNP.}
\item{rho_vv_H0}{The variant (homozygote) genotype probability with Allelic Bias = 0.5.}
\item{rho_rr_H0}{The reference (homozygote) genotype probability with Allelic Bias = 0.5.}
\item{rho_rv_H0}{The heterozyous genotype probability with Allelic Bias = 0.5.}
\item{data_hash}{Data frame. Input data frame with extra columns: allelefreq, genotype and genotype probabilities (prr, prv, pvv) per sample.}
}
\description{
\code{EMfit_binom} estimates, per locus, parameters of an assumed binomial mixture model via expectation maximization (genotype
frequencies, allelic bias i.e. heterozygous p-parameter).
Through this fit, per-samples genotype probabilities are obtained as well. A fit assuming no allelic bias (heterozygous p-parameter = 0.5) is
performed as well, for the purpose of significant allelic bias detection via likelihood ratio test.
}
