% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators.R
\name{fast}
\alias{fast}
\title{Speed up a pattern}
\usage{
fast(pat, factor)
}
\arguments{
\item{pat}{A pattern.}

\item{factor}{Speed factor.}
}
\value{
A pattern.
}
\description{
Play the pattern `factor` times faster (more events per cycle).
}
\examples{
query_arc(fast(s("bd"), 2), 0, 1)
}
