% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{from_list}
\alias{from_list}
\title{Cycle through a list of values}
\usage{
from_list(values)
}
\arguments{
\item{values}{A list of hap value lists, one per step.}
}
\value{
A pattern.
}
\description{
Each cycle is split into equal steps, one per list element, in order.
}
\examples{
query_arc(from_list(list(list(sound = "bd"), list(sound = "sd"))), 0, 1)
}
