% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators.R
\name{gain}
\alias{gain}
\title{Scale event gains}
\usage{
gain(pat, x)
}
\arguments{
\item{pat}{A pattern.}

\item{x}{Gain multiplier.}
}
\value{
A pattern.
}
\description{
Multiply every event gain by `x`.
}
\examples{
s("bd*4") |> gain(0.8)
}
