% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOF_uniform.R
\name{GOF_uniform}
\alias{GOF_uniform}
\title{Goodness-of-fit test comparing data following a uniform distribution to a Hardy-Weinberg distribution.}
\usage{
GOF_uniform(ref_counts, var_counts, allelefreq, SE, inbr = 0)
}
\arguments{
\item{ref_counts}{Numeric list. Reference counts.}

\item{var_counts}{Numeric list. Variant counts.}

\item{allelefreq}{Number. Allele frequency.}

\item{SE}{Number. Sequencing error rate.}

\item{inbr}{Number. Degree of inbreeding (default = 0).}
}
\value{
Geometric mean of the ratios of the putative heterozygous samples as a measure of the GOF.
}
\description{
\code{GOF_uniform} determines the GOF of the heterozygous samples to the HWE model.
}
\examples{
GOF_uniform(c(5, 8, 10, 3, 5, 6, 23), c(8, 8, 6, 4, 4, 10, 0), 0.5, 0.002, 0.12)
GOF_uniform(c(5, 0, 0, 3, 5, 1, 23), c(1, 8, 6, 2, 0, 10, 0), 0.5, 0.002)
}
