% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_pi_old}
\alias{grad_pi_old}
\title{Exact gradient of the beta-binomial log-likelihood function for pi using sums.}
\usage{
grad_pi_old(ms, ns, pi, theta)
}
\arguments{
\item{ms}{Numeric vector. Vector of number of successes}

\item{ns}{Numeric vector. Vector of number of trials}

\item{pi}{Number. Probability of success; \code{0 >= pi >= 1}}

\item{theta}{Number. Dispersion parameter; \code{0 >= theta > +Inf}}
}
\value{
A numeric vector of the same length as ms and ns, containing the gradient to pi in the give data points
}
\description{
\code{grad_pi_old} calculates the value of the gradient of the beta-binomial log-likelihood function to pi
at given data points via a number of sums, which is slow for high-value data
but fast for low-value data.
This function is called by \code{grad_pi} if necessary, and should not be called outside of this.
}
