% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_pi_pi}
\alias{grad_pi_pi}
\title{Beta-binomial second-order derivative with respect to its pi parameter (twice)}
\usage{
grad_pi_pi(ms, ns, pi, theta, MemLim = 2048L, Xtra = 7L)
}
\description{
Beta-binomial second-order derivative with respect to its pi parameter (twice)
}
