% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_theta}
\alias{grad_theta}
\title{Beta-binomial gradient with respect to its theta parameter}
\usage{
grad_theta(ms, ns, pi, theta, MemLim = 2048L, Xtra = 7L)
}
\description{
Beta-binomial gradient with respect to its theta parameter
}
