% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_theta_MP}
\alias{grad_theta_MP}
\title{Exact gradient of the beta-binomial log-likelihood function for theta using a multiprecision library.}
\usage{
grad_theta_MP(ms, ns, piX, thetaX, NecPres)
}
\arguments{
\item{ms}{Numeric vector. Vector of number of successes}

\item{ns}{Numeric vector. Vector of number of trials}

\item{piX}{Number. Probability of success; \code{0 >= piX >= 1}}

\item{thetaX}{Number. Dispersion parameter; \code{0 >= thetaX > +Inf}}

\item{NecPres}{Number. Necessary Precision, i.e. number of bits, for an accurate gradient calculatation, as determined by the function \code{grad_pi}}
}
\value{
A numeric vector of the same length as ms and ns, containing the gradient to theta in the give data points
}
\description{
\code{grad_theta_MP} calculates the value of the gradient of the beta-binomial log-likelihood function to theta
at given data points, while avoiding numerical mistakes (catastrophic cancellations) due to 
extreme parameter values. This function is called by \code{grad_theta}
if necessary, and should not be called outside of this.
}
