% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_theta_theta}
\alias{grad_theta_theta}
\title{Beta-binomial second-order derivative with respect to its theta-parameter (twice)}
\usage{
grad_theta_theta(ms, ns, pi, theta, MemLim = 2048L, Xtra = 7L)
}
\description{
Beta-binomial second-order derivative with respect to its theta-parameter (twice)
}
