% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GradPi_cppi}
\alias{GradPi_cppi}
\title{Beta-binomial gradient with respect to its pi parameter; for internal use by C++}
\usage{
GradPi_cppi(M, N, PI, THETA, MemLim, Xtra)
}
\description{
Beta-binomial gradient with respect to its pi parameter; for internal use by C++
}
