% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GradPiPi_cppi}
\alias{GradPiPi_cppi}
\title{Beta-binomial second-order derivative with respect to its pi parameter (twice); for internal use by C++}
\usage{
GradPiPi_cppi(M, N, PI, THETA, MemLim, Xtra)
}
\description{
Beta-binomial second-order derivative with respect to its pi parameter (twice); for internal use by C++
}
