% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GradPiTheta_cppi}
\alias{GradPiTheta_cppi}
\title{Beta-binomial second-order derivative with respect to its pi parameter (once) and theta-parameter (once); for internal use by C++}
\usage{
GradPiTheta_cppi(M, N, PI, THETA, MemLim, Xtra)
}
\description{
Beta-binomial second-order derivative with respect to its pi parameter (once) and theta-parameter (once); for internal use by C++
}
