% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GradTheta_cppi}
\alias{GradTheta_cppi}
\title{Beta-binomial gradient with respect to its theta parameter; for internal use by C++}
\usage{
GradTheta_cppi(M, N, PI, THETA, MemLim, Xtra)
}
\description{
Beta-binomial gradient with respect to its theta parameter; for internal use by C++
}
