% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GradThetaTheta_cppi}
\alias{GradThetaTheta_cppi}
\title{Beta-binomial second-order derivative with respect to its theta-parameter (twice); for internal use by C++}
\usage{
GradThetaTheta_cppi(M, N, PI, THETA, MemLim, Xtra)
}
\description{
Beta-binomial second-order derivative with respect to its theta-parameter (twice); for internal use by C++
}
