% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{hap_onset}
\alias{hap_onset}
\title{Onset of a hap}
\usage{
hap_onset(h)
}
\arguments{
\item{h}{A hap.}
}
\value{
A numeric onset in cycles.
}
\description{
Start of the hap in cycles: the start of `whole` when present, else the start of `part`.
}
\examples{
hap_onset(hap(timespan(0.5, 1), timespan(0.5, 1), list()))
}
