% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{highpass}
\alias{highpass}
\title{Highpass filter}
\usage{
highpass(x, cutoff, sr = SR)
}
\arguments{
\item{x}{Input signal.}

\item{cutoff}{Cutoff in Hz.}

\item{sr}{Sample rate, default 44100.}
}
\value{
The filtered signal.
}
\description{
Highpass via lowpass subtraction: x - lowpass(x).
}
\examples{
highpass(c(1, rep(0, 99)), 1000)
}
