% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWE_chisquared.R
\name{HWE_chisquared}
\alias{HWE_chisquared}
\title{Chi-squared test assessing the Hardy-Weinberg assumption.}
\usage{
HWE_chisquared(data, inbr = 0, results = NULL)
}
\arguments{
\item{data}{List or dataframe. A named list of per-SNP data frames (or one such dataframe) for a single chromosome, with names
corresponding to the position on the chromosome. The list contains genotype probabilities
("prr", "prv" and "pvv") for each sample ("sample").}

\item{inbr}{Number. Estimated or assumed degree of inbreeding within the population
under study (default = 0).}

\item{results}{Data frame. Data frame with SNP results in each row, containing
at least the chromosomal position of the SNP (results$position). If NULL, one dataframe is expected as the \code{data} input argument.}
}
\value{
The \code{results} input data frame appended with the following columns:
\item{HWEpval}{The p-value of the chi-squared test.}
\item{HWE_TestStat}{The test statistic of the chi-squared test.}
}
\description{
\code{HWE_chisquared} performs a per-SNP chi-squared test comparing observed genotype
   frequencies to those expected when assuming Hardy-Weinberg Equilibrium (HWE).
   A per-SNP corrected inbreeding coefficient is used if \code{Fmedian} is lower than the
   mathematically minimal possible value of -q/(1-q), with q the minor allele frequency (i.e. this minimum is used instead).
}
