% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impr_LOI_analysis.R
\name{impr_LOI_analysis}
\alias{impr_LOI_analysis}
\title{multi-locus wrapper function for imprinting detection and loss-of-imprinting analysis}
\usage{
impr_LOI_analysis(
  DataList,
  SE,
  inbr,
  MinMinorAllelefreq = 0.15,
  sym_filt = 0.05,
  adj_p_filt = 0.05,
  gof_filt = 0.8,
  med_i_filt = 0.8,
  i_filt = 0.6
)
}
\arguments{
\item{DataList}{A list containing two lists of dataframes, which in turn contain control- and case-data (in that order). Subsequent entries in both lists
correspond to subsequent loci (lists must be the same size and named per-locus). Each of the dataframes should at least contain both a \code{ref_count} and
\code{var_count} column; control dataframes should also contain the columns \code{allelefreq_prel} (from running \code{AllelicMeta_est}), \code{locus_id}, 
\code{ref}, \code{var}, and \code{coverage} (from running \code{prior_filter}).}

\item{SE}{Number. Sequencing error rate, population metaparameter.}

\item{inbr}{Number. Inbreeding coefficient, population metaparameter.}

\item{MinMinorAllelefreq}{Number. Minimal allowed minor allele frequency as determinde via the estimated reference allele frequency by \code{AllelicMeta_est} to consider the locus for imprinting analysis.}

\item{sym_filt}{Number. Minimum required symmetry statistic to consider the locus for imprinting analysis, see \code{symmetry_gof}.}

\item{adj_p_filt}{Number. Required FDR-corrected significance level of imprinting for loci to be considered for loss-of-imprinting detection, see the function \code{final_filter}.}

\item{gof_filt}{Number. Required minimum goodness-of-fit heuristic of imprinted loci to be considered for loss-of-imprinting detection, see the function \code{final_filter}.}

\item{med_i_filt}{Number. Required minimum median imprinting of imprinted loci to be considered for loss-of-imprinting detection, see the functions \code{median_imprinting} and \code{final_filter}.}

\item{i_filt}{Number. Required minimum estimated imprinting of imprinted loci to be considered for loss-of-imprinting detection, see the functions \code{imprinting_est} and \code{final_filter}.}
}
\value{
A list containing the following components:
\item{impr_res}{Results of the imprinting detection analysis, containing, for every locus:}
\itemize{
     \item{position}{The locus' name, according to \code{names(DataList)}}
     \item{LRT}{The test statistic of the likelihood ratio test against no imprinting (see \code{imprinting_est}).}
     \item{p}{The p-value of the likelihood ratio test against no imprinting (see \code{imprinting_est}).}
     \item{estimated.i}{The estimated degree of imprinting (see \code{imprinting_est}).}
     \item{allele.frequency}{Reference allele frequency as estimated by \code{AllelicMeta_est}, already given as input.}
     \item{reference}{Reference allele nucleotide.}
     \item{variant}{Variant allele nucleotide.}
     \item{med_cov}{Median coverage (reference + variant) across the locus.}
     \item{nr_samples}{Number of samples (with a total reference + variant read count of at least 1) covering the locus.}
     \item{GOF}{Goodness-of-fit heuristic as determined by \code{imprinting_est}.}
     \item{symmetry}{Symmetry statistic as determined by \code{symmetry_gof}.}
     \item{med_impr}{Median imprinting as determined by \code{median_imprinting.}}
   }
\item{LOI_res}{A dataframe containing, only for loci fitting well to maelstRom's imprinting model (see \code{imprinting_est}, determined by \code{gof_filt}) 
that are significantly (determined by the \code{adj_p_filt input}) and suffiently (determined by \code{i_filt} and \code{med_i_filt}) imprinted,
the results of the differential imprinting analysis (which equals loss of imprinting if there's less imprinting in cases). This contains the same
columns as \code{impr_res}, but supplemented with:}
   \itemize{
     \item{DI_pval}{The p-value testing for differential imprinting using logistic regression, see the documentation for \code{LOItest_logreg} for more details.}
   }
}
\description{
\code{impr_LOI_analysis} is a wrapper function of \code{symmetry_gof}, \code{imprinting_est}, \code{median_imprinting}, \code{final_filter}, and
\code{LOItest_logreg} performing maelstRom's entire (loss of) imprinting pipeline. It goes over the following steps:
\enumerate{
  \item Prior filtering on basis of the \code{allelefreq_prel} column that should be present in the control dataframe,
  which it should be after running \code{AllelicMeta_est}.
  \item Prior filtering via \code{symmetry_gof}, see its documentation for more details.
  \item Running \code{imprinting_est} to detect imprinted loci and write results to \code{impr_res}; this includes a measure of median
  imprinting as returned by the function \code{median_imprinting} as well.
  \item Filter \code{impr_res} down to significantly and sufficiently imprinted loci using \code{final_filter}, then perform loss-of-imprinting
  detection using \code{LOItest_logreg}. LOI-detection results are written to \code{LOI_res}.
}
}
