% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprinting_est.R
\name{imprinting_est}
\alias{imprinting_est}
\title{Estimate the degree of imprinting.}
\usage{
imprinting_est(ref_counts, var_counts, allelefreq, SE, inbr = 0)
}
\arguments{
\item{ref_counts}{Numeric list. Reference counts.}

\item{var_counts}{Numeric list. Variant counts.}

\item{allelefreq}{Number. Allele frequency.}

\item{SE}{Number. Sequencing error rate.}

\item{inbr}{Number. Degree of inbreeding (default = 0).}
}
\value{
A list containing the following components:
\item{est_i}{The estimated degree of imprinting.}
\item{LRT}{The test statistic of the likelihood ratio test against no imprinting.}
\item{p_value}{The p-value of the likelihood ratio test against no imprinting.}
\item{GOF_likelihood}{A goodness-of-fit value based on count-corrected likelihood, i.e. mean log PMF-value according to \code{pmf_impr} multiplied by each sample's coverage + 1 across samples of a locus.}
}
\description{
\code{imprinting_est} estimates the degree of imprinting using iterative likelihood ratio tests for a specific SNP,
going from imprinting = 0 to imprinting = 1 in steps of 0.01, performing LRTs along the way (against imprinting = 0), and finally retaining the best fit.
See the function \code{pmf_impr} for what the degree of imprinting entails, mathematically
}
\examples{
imprinting_est(c(5, 8, 10, 3, 5, 6, 23), c(8, 8, 6, 4, 4, 10, 0), 0.5, 0.002)
imprinting_est(c(5, 0, 0, 3, 5, 1, 23), c(1, 8, 6, 2, 0, 10, 0), 0.5, 0.002, 0.12)
}
