% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini.R
\name{is_num}
\alias{is_num}
\title{Test whether a string is a number}
\usage{
is_num(s)
}
\arguments{
\item{s}{A character string.}
}
\value{
A logical value.
}
\description{
True when the string parses as numeric.
}
\examples{
is_num("1.5")
}
