% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{is_rest}
\alias{is_rest}
\title{Test whether a hap is a rest}
\usage{
is_rest(h)
}
\arguments{
\item{h}{A hap.}
}
\value{
A logical value.
}
\description{
A hap is a rest when its value is empty or has `rest = TRUE`.
}
\examples{
is_rest(hap(NULL, timespan(0, 1), list()))
}
