% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogGaussianQuad_CPP}
\alias{LogGaussianQuad_CPP}
\title{Internal helper function for TumPur_LogLik_CPP when NumIntMethod == "GaussianQuad"}
\usage{
LogGaussianQuad_CPP(lower, upper, a, b, q, TC, RC, curTR, curRC, Wvec, Nvec)
}
\description{
Internal helper function for TumPur_LogLik_CPP when NumIntMethod == "GaussianQuad"
}
