% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogNewtonCotes_CPP}
\alias{LogNewtonCotes_CPP}
\title{Internal helper function for TumPur_LogLik_CPP when NumIntMethod == "NewtonCotes"}
\usage{
LogNewtonCotes_CPP(lower, upper, a, b, q, TC, RC, curTR, curRC, Wvec)
}
\description{
Internal helper function for TumPur_LogLik_CPP when NumIntMethod == "NewtonCotes"
}
