% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogTanhSinhQuad_CPP}
\alias{LogTanhSinhQuad_CPP}
\title{Internal helper function for TumPur_LogLik_CPP when NumIntMethod == "TanhSinhQuad"}
\usage{
LogTanhSinhQuad_CPP(lower, upper, n, a, b, q, TC, RC, curTR, curRC, prec)
}
\description{
Internal helper function for TumPur_LogLik_CPP when NumIntMethod == "TanhSinhQuad"
}
