% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogTrapezoidalInt_CPP}
\alias{LogTrapezoidalInt_CPP}
\title{Internal helper function for TumPur_LogLik_CPP when NumIntMethod == "Gregory"}
\usage{
LogTrapezoidalInt_CPP(lower, upper, n, a, b, q, TC, RC, curTR, curRC)
}
\description{
Internal helper function for TumPur_LogLik_CPP when NumIntMethod == "Gregory"
}
