% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOI_calling.R
\name{LOI_calling}
\alias{LOI_calling}
\title{Call LOI samples over al imprinted loci.}
\usage{
LOI_calling(data_hash, samples_all, SE)
}
\arguments{
\item{data_hash}{Hash. Hash of SNP positions with a data frame for every position with at least columns of allelic
counts ("A", "T", "C", "G"), reference and variants counts ("ref_count" and "var_count", respectively) and
sample ids ("samples").}

\item{samples_all}{Character list. List of all sample names.}

\item{SE}{Number. Sequencing error rate.}
}
\value{
Data frame with LOI calls per samples. The data frame contains sample ids ("samples"), number of least and
   most expressed alleles ("least" and "most", respectively), a quality check ("warning"), binomial p-values ("binom"),
   FDR adjusted p-values ("FDR") and LOI vs not LOI calls ("geno"). A "!" in the warning column indicates that more
   non-standard alleles than standard alleles were present for a sample (with a count of at least 2).
}
\description{
\code{LOI_calling} calls LOI samples based on binomial p-value.
}
