% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOItest_logreg.R
\name{LOItest_logreg}
\alias{LOItest_logreg}
\title{Logistic regression on the degree of heterozygosity between cases and controls}
\usage{
LOItest_logreg(
  ref_counts_ctrl,
  var_counts_ctrl,
  ref_counts_case,
  var_counts_case
)
}
\arguments{
\item{ref_counts_ctrl, var_counts_ctrl}{Numeric vectors. Reference and variant counts of control samples.}

\item{ref_counts_case, var_counts_case}{Numeric vectors. Reference and variant counts of case samples.}
}
\value{
Beta-value (beta) and p-value (p.value) of the generalised linear model comparing the fraction of heterozygous samples
    in cases versus controls.
}
\description{
\code{LOItest_logreg} performs logistic regression on the degree of heterozygosity in cases compared to control samples.
    This tests whether there is a significant difference (p-value) between the fraction of heterozygous samples in
    cases versus controls, which (in the case of a higher fraction in cases) can indicate loss of imprinting.
}
\examples{
LOItest_logreg(c(5, 8, 10, 3, 5, 6, 23), c(8, 8, 6, 4, 4, 10, 0),
    c(5, 0, 0, 3, 5, 1, 23), c(1, 8, 6, 2, 0, 10, 0))
}
