% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maelstRom_ADChromplot.R
\name{maelstRom_ADChromplot}
\alias{maelstRom_ADChromplot}
\title{Make a plot summarizing allelic divergence (-related phenomena) across a chromosome}
\usage{
maelstRom_ADChromplot(
  AD_Data,
  DE_Data,
  Meth_Data = NULL,
  CNAgain_Data = NULL,
  CNAloss_Data = NULL,
  pvalSIG = 0.05,
  roll_median = 15
)
}
\arguments{
\item{AD_Data}{Dataframe. Should at least contain chromosomal position ("pos"), overdispersion in controls and cases expressed as rho which lies between
0 and 1 ("rho_control", "rho_case") and the p-value testing for differential allelic divergence (difference in overdispersion parameter between controls and cases; "pval")}

\item{DE_Data}{Dataframe. Should at least contain chromosomal position ("pos") and the log(2)FC of expression between cases and controls ("logFC")}

\item{Meth_Data}{Dataframe. Optional; should at least contain chromosomal position ("pos") and the p-value testing for case hypermethylation ("logFC")}

\item{CNAgain_Data}{Dataframe. Optional; should at least contain chromosomal position ("pos") and the average gain in cases ("CNAgain")}

\item{CNAloss_Data}{Dataframe. Optional; should at least contain chromosomal position ("pos") and the average loss in cases ("CNAloss")}

\item{pvalSIG}{Number. Significance cutoff of provided p-values (these should be FDR-corrected in advance)}

\item{roll_median}{integer. All measures are plotted as rolling medians across \code{roll_median} data points.}

\item{allelefreq}{Number. Allele frequency.}
}
\value{
A list containing the following components:
\item{ADDE_plot}{A plot vizualizing Allelic Divergence and Differential Expression results.}
\item{MethCNA_plot}{A plot vizualizing hypermethylation and copy-number-alteration results.}
\item{LEG1, LEG2, LEG3}{Various legend items, to be combined with \code{ADDE_plot} and \code{MethCNA_plot} if the user so desires, e.g. using the \code{patchwork} package.}
}
\description{
\code{maelstRom_ADChromplot} returns various ggplot-objects visualizing maelstRom's differential allelic divergence results across a chromosome,
together with differential expression logFC which can be determined using the same RNAseq data, as well as (optionally) methylation- and copy-number-alteration-data if available.
All of this data should be supplied as separate dataframes, which enables plotting even when some of this data is missing for some of the loci or genes.
}
