% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGE_EMfitplot.R
\name{MAGE_EMfitplot}
\alias{MAGE_EMfitplot}
\title{Plot MAGE's EM-fit results.}
\usage{
MAGE_EMfitplot(
  ref_counts,
  var_counts,
  pr,
  prv,
  pv,
  theta_hom,
  theta_het,
  pr_NoShift = NULL,
  prv_NoShift = NULL,
  pv_NoShift = NULL,
  theta_hom_NoShift = NULL,
  theta_het_NoShift = NULL,
  probshift,
  SE,
  MinCount = 0,
  ScaleCount = 50,
  ScaleHist = TRUE,
  nbins = 100,
  plot_NoShift = TRUE,
  SplitPeaks = TRUE,
  ShiftCols = NULL,
  NoShiftCol = NULL,
  wd_res = NULL,
  chr = "",
  position = "",
  gene = ""
)
}
\arguments{
\item{ref_counts}{Numeric vector. Reference counts.}

\item{var_counts}{Numeric vector. Variant counts.}

\item{pr}{Number. Reference homozygote genotype probability of the locus.}

\item{prv}{Number. Heterozygote genotype probability of the locus.}

\item{pv}{Number. Variant homozygote genotype probability of the locus.}

\item{theta_hom}{Number. The dispersion parameter of the homozygous peaks.}

\item{theta_het}{Number. The dispersion parameter of the heterozygous peak.}

\item{pr_NoShift}{Number. Reference homozygote genotype probability of the locus under the null hypothesis fit (no allelic bias; heterozygous pi-parameter = 0.5).}

\item{prv_NoShift}{Number. Heterozygote genotype probability of the locus under the null hypothesis fit (no allelic bias; heterozygous pi-parameter = 0.5).}

\item{pv_NoShift}{Number. Variant homozygote genotype probability of the locus under the null hypothesis fit (no allelic bias; heterozygous pi-parameter = 0.5).}

\item{theta_hom_NoShift}{Number. The dispersion parameter of the homozygous peaks under the null hypothesis fit (no allelic bias; heterozygous pi-parameter = 0.5).}

\item{theta_het_NoShift}{Number. The dispersion parameter of the heterozygous peak under the null hypothesis fit (no allelic bias; heterozygous pi-parameter = 0.5).}

\item{probshift}{Number. The reference allele fraction in heterozygotes, indicating allelic bias when deviating from 0.5}

\item{SE}{Number. Sequencing error rate.}

\item{MinCount}{Number. A minimal count filter for plotting only; samples with a lower count will not be included in the plotted histogram.}

\item{ScaleCount}{Number. Even though the observed number of reads can vary per sample, we have to assume a single number of observed reads to plot PMFs. One can opt to pick
a set value for this (the default is 50) or e.g. use the locus' median or mean coverage here, so as to best correspond to the observed data.}

\item{ScaleHist}{Logical. See the remark given at \code{ScaleCount}; if TRUE, tranforms the input data to all have the same total allele counts \code{ScaleCount}, with the actual number
of reference reads being determined as the quantile function assuming \code{ScaleCount} total counts with the same probability as the observed data (this probability being calculated
with the actually observed number of counts). The idea here is to better reflect how well the observed data corresponds to the fitted PMF, which is plotted for a single number of observed reads.}

\item{nbins}{Number. Number of bins for the histogram depicting the observed reference allele fractions.}

\item{plot_NoShift}{Logical. If TRUE, also plots the PMF of the unshifted fit (no allelic bias; heterozygous pi-parameter = 0.5)}

\item{SplitPeaks}{Logical. If TRUE, the PMFs of both homozygotes and the heterozygotes are plotted seperately and using different colors. If FALSE, one total PMF of the mixture model is plotted.}

\item{ShiftCols}{String (vector). Colors for the plotted PMFs of the fit incorporating allelic bias. Three values are required if \code{SplitPeaks} is TRUE, otherwise just one value is required.
Default colors are used in case none are given as input.}

\item{NoShiftCol}{String. Color for the unshifted PMF (no allelic bias; heterozygous pi-parameter = 0.5). There's no option to split the peaks of this unshifted fit as it would make
the final figure way too clustered, so only one color value is required. A default color is used in case non is given as input.}

\item{wd_res}{String. Working directory where plots are saved; if non is given, the plot itself (as a \code{ggplot2} object) is returned instead.}

\item{chr}{Number. Chromosome of the locus, to be used in naming the outpud file if a \code{wd_res} is given.}

\item{position}{Number. Position of the locus, to be used in naming the outpud file if a \code{wd_res} is given.}

\item{gene}{String. Gene the locus is part of, to be used in naming the outpud file if a \code{wd_res} is given; optional.}
}
\description{
\code{MAGE_EMfitplot} plots the results of MAGE's beta-binomial EM-fit. More specifically, it plots both the observed reference allele fractions as a histogram (optionally scaled
to a single total count parameter, see \code{ScaleHist} and \code{ScaleCount}) and the parameters yielded by the EM-fit as beta-binomial PMFs (assuming a total read counts \code{ScaleCount}
and compressed to the zero-to-one interval of the reference allele fraction instead of integer reference counts). Optionally, the unshifted fit assuming no allelic bias (heterozygous
pi-parameter = 0.5) is plotted as well, if \code{plot_NoShift} is TRUE and its distributional parameters are provided as input.
}
