% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_AB.R
\name{median_AB}
\alias{median_AB}
\title{Calculates the median allelic bias}
\usage{
median_AB(ref_counts, var_counts, allelefreq, inbr = 0)
}
\arguments{
\item{ref_counts}{Numeric vector. Reference counts.}

\item{var_counts}{Numeric vector. Variant counts.}

\item{allelefreq}{Number. Allele frequency.}

\item{inbr}{Number. Inbreeding coefficient (default = 0).}
}
\value{
Median heterozygous allelic bias of \code{ref_counts} and \code{var_counts} taking into account
\code{allelefreq} and \code{inbr}.
}
\description{
\code{median_AB} calculates median allelic bias of a given locus, i.e. the heterozygous reference allele ratio.
It does this by ordering the samples according to degree of heterozygosity (i.e. minor over major allele count), 
then calculating the location of the "median heterozyogous sample" assuming HWE given input parameters
(\code{allelefreq}, \code{inbr}) as \code{round(number of samples \* allelefreq \* (1 - allelefreq) \* (1 - inbr))}, 
2 times this unrounded number being the expected number of heterozygotes. Of this median heterozygous observation,
the allelic bias is returned (reference over total allele ratio).
}
\examples{
median_AB(c(5, 8, 10, 3, 5, 6, 23), c(8, 8, 6, 4, 4, 10, 0), 0.5, 0.12)
median_AB(c(5, 0, 0, 3, 5, 1, 23), c(1, 8, 6, 2, 0, 10, 0), 0.5, 0.12)
median_AB(c(5, 0, 0, 3, 5, 1, 23), c(1, 8, 6, 2, 0, 10, 0), 0.5)
}
