% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_imprinting.R
\name{median_imprinting}
\alias{median_imprinting}
\title{Calculate median imprinting.}
\usage{
median_imprinting(ref_counts, var_counts, allelefreq, inbr = 0)
}
\arguments{
\item{ref_counts}{Numeric vector. Reference counts.}

\item{var_counts}{Numeric vector. Variant counts.}

\item{allelefreq}{Number. Allele frequency.}

\item{inbr}{Number. Inbreeding coefficient (default = 0).}
}
\value{
Median degree of imprinting of \code{ref_counts} and \code{var_counts} taking into account \code{allelefreq} and \code{inbr}.
}
\description{
\code{median_imprinting} calculates median imprinting of allelic counts for a locus. 
It does this by ordering the samples according to degree of heterozygosity (i.e. minor over major allele count), 
then calculating the location of the "median heterozyogous sample" assuming HWE given input parameters
(\code{allelefreq}, \code{inbr}) as \code{round(number of samples \* allelefreq \* (1 - allelefreq) \* (1 - inbr))}, 
2 times this unrounded number being the expected number of heterozygotes. Of this median heterozygous observation,
the degree of imprining is returned as \code{2 \* (0.5 - minor over major allele count)}.
}
\examples{
median_imprinting(c(5, 8, 10, 3, 5, 6, 23), c(8, 8, 6, 4, 4, 10, 0), 0.5, 0.12)
median_imprinting(c(5, 0, 0, 3, 5, 1, 23), c(1, 8, 6, 2, 0, 10, 0), 0.5, 0.12)
median_imprinting(c(5, 0, 0, 3, 5, 1, 23), c(1, 8, 6, 2, 0, 10, 0), 0.5)
}
