% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{midi_to_freq}
\alias{midi_to_freq}
\title{MIDI number to frequency}
\usage{
midi_to_freq(midi)
}
\arguments{
\item{midi}{A MIDI number.}
}
\value{
A frequency in Hz.
}
\description{
Equal-tempered frequency in Hz (A4 = 440).
}
\examples{
midi_to_freq(69)
}
