% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators.R
\name{midi_to_note}
\alias{midi_to_note}
\title{MIDI number to note name}
\usage{
midi_to_note(midi)
}
\arguments{
\item{midi}{A MIDI number.}
}
\value{
A note name like "c4".
}
\description{
Inverse of `note_to_midi`: 69 becomes "a4". Uses sharps.
}
\examples{
midi_to_note(69)
}
