% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini.R
\name{mini_compile}
\alias{mini_compile}
\title{Compile an AST node to a pattern}
\usage{
mini_compile(node, span, mode)
}
\arguments{
\item{node}{An AST node.}

\item{span}{A TimeSpan.}

\item{mode}{One of "sound", "note", "n", or "token".}
}
\value{
A pattern.
}
\description{
Recursively compile a parsed AST node over `span` in the given mode.
}
\examples{
query_arc(mini_compile(mini_parse_root("bd"), timespan(0, 1), "sound"), 0, 1)
}
