% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini.R
\name{mini_parse_choice}
\alias{mini_parse_choice}
\title{Parse a random choice}
\usage{
mini_parse_choice(ps)
}
\arguments{
\item{ps}{A parser.}
}
\value{
An AST node.
}
\description{
Parse `a | b | c` into a choice node (one item per cycle).
}
\examples{
mini_parse_choice(new_parser(mini_tokenize("bd|sd")))
}
