% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini.R
\name{mini_parse_element}
\alias{mini_parse_element}
\title{Parse an element with suffixes}
\usage{
mini_parse_element(ps)
}
\arguments{
\item{ps}{A parser.}
}
\value{
An AST node.
}
\description{
Parse a primary, then bind `*`, `/`, `?`, `:`, and `(,)` suffixes tighter than `|`.
}
\examples{
mini_parse_element(new_parser(mini_tokenize("bd*4")))
}
