% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini.R
\name{mini_parse_layer}
\alias{mini_parse_layer}
\title{Parse one mini-notation layer}
\usage{
mini_parse_layer(ps)
}
\arguments{
\item{ps}{A parser.}
}
\value{
A `seq` AST node.
}
\description{
Parse a sequence of steps, expanding `!n` replication and `a .. b` ranges.
}
\examples{
mini_parse_layer(new_parser(mini_tokenize("bd sd")))
}
