% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini.R
\name{mini_parse_number}
\alias{mini_parse_number}
\title{Parse a numeric token}
\usage{
mini_parse_number(ps)
}
\arguments{
\item{ps}{A parser.}
}
\value{
A numeric value.
}
\description{
Consume one word token that parses as a number.
}
\examples{
ps <- new_parser(mini_tokenize("4")); mini_parse_number(ps)
}
