% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini.R
\name{mini_parse_pattern}
\alias{mini_parse_pattern}
\title{Parse comma-separated layers}
\usage{
mini_parse_pattern(ps)
}
\arguments{
\item{ps}{A parser.}
}
\value{
An AST node (a `par` node when there are several layers).
}
\description{
Parse parallel layers separated by commas into one AST node.
}
\examples{
mini_parse_pattern(new_parser(mini_tokenize("bd, hh*8")))
}
