% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini.R
\name{mini_parse_primary}
\alias{mini_parse_primary}
\title{Parse a primary}
\usage{
mini_parse_primary(ps)
}
\arguments{
\item{ps}{A parser.}
}
\value{
An AST node.
}
\description{
Parse a word atom (including the `6%5` ratio shorthand), a `[...]` group, a `<...>` alternation, or a `{...}` polymeter.
}
\examples{
mini_parse_primary(new_parser(mini_tokenize("bd")))
}
