% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{mix}
\alias{mix}
\title{Add two signals}
\usage{
mix(a, b)
}
\arguments{
\item{a}{A numeric vector.}

\item{b}{A numeric vector.}
}
\value{
The summed vector.
}
\description{
Sum two vectors, padding the shorter with zeros.
}
\examples{
mix(c(1, 2), c(1))
}
