% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MomentEst_MixedBetaBinom_SESlack.R
\name{MomentEst_MixedBetaBinom_SESlack}
\alias{MomentEst_MixedBetaBinom_SESlack}
\title{This function is a work in progress. As such, it is not exported yet; but it allows for some leniency on the sequencing error parameter, hence "SESlack"}
\usage{
MomentEst_MixedBetaBinom_SESlack(
  ref_counts,
  var_counts,
  spr,
  spv,
  sprv,
  pi_RR_fix = NULL,
  pi_VV_fix = NULL,
  pi_het_fix = NULL
)
}
\description{
This function is a work in progress. As such, it is not exported yet; but it allows for some leniency on the sequencing error parameter, hence "SESlack"
}
