% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MPhelper_GradThetaTheta}
\alias{MPhelper_GradThetaTheta}
\title{Helper function for GradThetaTheta_cppi}
\usage{
MPhelper_GradThetaTheta(M, N, PI, NecPres)
}
\description{
Helper function for GradThetaTheta_cppi
}
