% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_coeff.R
\name{multinomial_coeff}
\alias{multinomial_coeff}
\title{Calculate multinomial coefficient(s).}
\usage{
multinomial_coeff(ref_counts, var_counts)
}
\arguments{
\item{ref_counts}{Numeric vector. Reference count(s).}

\item{var_counts}{Numeric vector. Variant count(s).}
}
\value{
The multinomial coefficient(s) of \code{ref_counts} and \code{var_counts}.
}
\description{
\code{multinomial_coeff} calculates multinomial coefficients given locus data, mainly for internal use by \code{imprinting_est}.
}
\examples{
multinomial_coeff(5, 11)
multinomial_coeff(c(5, 8, 10, 3, 5, 6, 23), c(8, 8, 6, 4, 4, 10, 0))
}
