% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini.R
\name{note}
\alias{note}
\title{Mini-notation for pitches}
\usage{
note(text)
}
\arguments{
\item{text}{Mini-notation text.}
}
\value{
A pattern.
}
\description{
Parse text into a pattern of `list(note = ...)` values.
}
\examples{
query_arc(note("c4 eb4 g4"), 0, 1)
}
