% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{note_to_midi}
\alias{note_to_midi}
\title{Note name to MIDI number}
\usage{
note_to_midi(name)
}
\arguments{
\item{name}{A note name.}
}
\value{
An integer MIDI number.
}
\description{
Parse names like "a4", "c#3", "eb2" ("s" also means sharp, "b" means flat; default octave 3).
}
\examples{
note_to_midi("a4")
}
