% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators.R
\name{often}
\alias{often}
\title{Apply a transform on most cycles}
\usage{
often(pat, fn)
}
\arguments{
\item{pat}{A pattern.}

\item{fn}{A function from pattern to pattern.}
}
\value{
A pattern.
}
\description{
`sometimes_by` with probability 0.75.
}
\examples{
query_arc(often(s("bd*4"), function(p) rev(p)), 0, 2)
}
