% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators.R
\name{palindrome}
\alias{palindrome}
\title{Alternate forward and reversed cycles}
\usage{
palindrome(pat)
}
\arguments{
\item{pat}{A pattern.}
}
\value{
A pattern.
}
\description{
Even cycles play forward, odd cycles play reversed.
}
\examples{
query_arc(palindrome(s("bd sd")), 0, 2)
}
