% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators.R
\name{pan}
\alias{pan}
\title{Stereo pan}
\usage{
pan(pat, x)
}
\arguments{
\item{pat}{A pattern.}

\item{x}{Pan position in [0, 1].}
}
\value{
A pattern.
}
\description{
Pan position per event: 0 is hard left, 1 is hard right, 0.5 is center.
}
\examples{
s("hh*8") |> pan(0.2)
}
