% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{pat_filter}
\alias{pat_filter}
\title{Filter haps by predicate}
\usage{
pat_filter(p, pred)
}
\arguments{
\item{p}{A pattern.}

\item{pred}{A function from a hap to a logical value.}
}
\value{
A pattern.
}
\description{
Keep only the haps for which `pred` returns TRUE.
}
\examples{
query_arc(pat_filter(s("bd*4"), function(h) TRUE), 0, 1)
}
