% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{pat_overlay}
\alias{pat_overlay}
\title{Layer two patterns}
\usage{
pat_overlay(a, b)
}
\arguments{
\item{a}{A pattern.}

\item{b}{A pattern.}
}
\value{
A pattern.
}
\description{
Combine two patterns so the result sounds the haps of both.
}
\examples{
query_arc(pat_overlay(s("bd"), s("sd")), 0, 1)
}
