% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{pat_with_value}
\alias{pat_with_value}
\title{Transform hap values}
\usage{
pat_with_value(p, fn)
}
\arguments{
\item{p}{A pattern.}

\item{fn}{A function from a value list to a value list.}
}
\value{
A pattern.
}
\description{
Apply a function to every hap value in a pattern.
}
\examples{
query_arc(pat_with_value(s("bd"), function(v) { v$gain <- 0.5; v }), 0, 1)
}
