% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators.R
\name{per_cycle}
\alias{per_cycle}
\title{Apply a function per cycle}
\usage{
per_cycle(pat, fn)
}
\arguments{
\item{pat}{A pattern.}

\item{fn}{A function of (pattern, TimeSpan, cycle number).}
}
\value{
A pattern.
}
\description{
Split the query into whole cycles and call `fn(pat, subspan, cycle)` for each.
}
\examples{
query_arc(per_cycle(s("bd*4"), function(p, sub, cyc) p(sub)), 0, 2)
}
