% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_betabinomMix_SEslack.R
\name{pmf_betabinomMix_SEslack}
\alias{pmf_betabinomMix_SEslack}
\title{This function is a work in progress. As such, it is not exported yet; but it allows for some leniency on the sequencing error parameter, hence "SESlack"}
\usage{
pmf_betabinomMix_SEslack(
  ref_counts,
  var_counts,
  probshift,
  piRR,
  piVV,
  pr,
  pv,
  prv,
  theta_RR,
  theta_VV,
  theta_het
)
}
\description{
This function is a work in progress. As such, it is not exported yet; but it allows for some leniency on the sequencing error parameter, hence "SESlack"
}
