% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf_impr.R
\name{pmf_impr}
\alias{pmf_impr}
\title{Probability Mass Function for an imprinted SNP.}
\usage{
pmf_impr(ref_counts, var_counts, allelefreq, impr, SE, inbr = 0)
}
\arguments{
\item{ref_counts}{Number or Numeric vector Reference count(s).}

\item{var_counts}{Number or Numeric vector. Variant count(s).}

\item{allelefreq}{Number. Allele frequency.}

\item{impr}{Number. Degree of imprinting.}

\item{SE}{Number. Sequencing error rate.}

\item{inbr}{Number. Degree of inbreeding (default = 0).}
}
\value{
Probability of observing \code{ref_counts} and \code{var_counts}
}
\description{
\code{pmf_impr} calculates the probability of observing population-level RNAseq data (reference- and variant allele counts per sample)
assuming an imprinted binomial mixture model with the mean reference allele fraction in heterozygotes fixed at 0.5 (no allelic bias).
More specifically, the formula used is (though under-the-hood not using R's built-in dbinom function but our own mathematical implementation):

pr * dbinom(x = ref_counts, size = ref_counts + var_counts, prob = 1-SE) +

0.5 * prv * dbinom(x = ref_counts, size = ref_counts + var_counts, prob = (0.5-(impr/2)) / (1-(impr/2)) * (1-SE) + (0.5/1-(impr/2)) * SE) +

0.5 * prv * dbinom(x = ref_counts, size = ref_counts + var_counts, prob = (0.5-(impr/2)) / (1-(impr/2)) * SE + (0.5/1-(impr/2)) * (1-SE)) +

pv * dbinom(x = ref_counts, size = ref_counts + var_counts, prob = SE)

pr, pv and prv are calculated assuming Hardy-Weinberg-Equilibrium with given input parameters (inbreeding coefficient \code{inbr} and reference allele frequency \code{allelefreq})
}
\examples{
pmf_impr(10, 10, 0.5, 0, 0.002, 0.12)
pmf_impr(10, 10, 0.5, 1, 0.002)
pmf_impr(0, 10, 0.5, 1, 0.002, 0.12)
pmf_impr(c(5, 8, 10, 3, 5, 6, 23), c(8, 8, 6, 4, 4, 10, 0), 0.5, 1, 0.002, 0.12)
pmf_impr(c(5, 0, 0, 3, 5, 1, 23), c(1, 8, 6, 2, 0, 10, 0), 0.5, 1, 0.002)
}
