% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini.R
\name{ps_is_op}
\alias{ps_is_op}
\title{Test for an operator token}
\usage{
ps_is_op(ps, op)
}
\arguments{
\item{ps}{A parser.}

\item{op}{An operator character like "*" or "?".}
}
\value{
A logical value.
}
\description{
True when the current token is the operator `op`.
}
\examples{
ps <- new_parser(mini_tokenize("bd*4")); ps_next(ps); ps_is_op(ps, "*")
}
