% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini.R
\name{ps_next}
\alias{ps_next}
\title{Consume the current token}
\usage{
ps_next(ps)
}
\arguments{
\item{ps}{A parser.}
}
\value{
The consumed token.
}
\description{
Return the token under the cursor and advance past it.
}
\examples{
ps <- new_parser(mini_tokenize("bd")); ps_next(ps)
}
