% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini.R
\name{ps_peek_val}
\alias{ps_peek_val}
\title{Value of the current token}
\usage{
ps_peek_val(ps)
}
\arguments{
\item{ps}{A parser.}
}
\value{
The token value or NULL.
}
\description{
Shorthand for the value part of `ps_peek(ps)`.
}
\examples{
ps <- new_parser(mini_tokenize("bd")); ps_peek_val(ps)
}
