% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{query_arc}
\alias{query_arc}
\title{Query the haps beginning in a time range}
\usage{
query_arc(p, start, cycles)
}
\arguments{
\item{p}{A pattern.}

\item{start}{Start in cycles.}

\item{cycles}{Number of cycles to query.}
}
\value{
A list of haps, sorted by onset.
}
\description{
Run a pattern over [start, start + cycles) and return the haps whose onset falls inside, sorted by onset.
}
\examples{
query_arc(s("bd*4"), 0, 1)
}
