% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators.R
\name{rarely}
\alias{rarely}
\title{Apply a transform on few cycles}
\usage{
rarely(pat, fn)
}
\arguments{
\item{pat}{A pattern.}

\item{fn}{A function from pattern to pattern.}
}
\value{
A pattern.
}
\description{
`sometimes_by` with probability 0.25.
}
\examples{
query_arc(rarely(s("bd*4"), function(p) rev(p)), 0, 4)
}
