% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{render}
\alias{render}
\title{Render a pattern to mono samples}
\usage{
render(pat, cycles, cps = 0.5, sr = SR)
}
\arguments{
\item{pat}{A pattern.}

\item{cycles}{Number of cycles to render.}

\item{cps}{Cycles per second, default 0.5.}

\item{sr}{Sample rate, default 44100.}
}
\value{
A numeric vector of samples.
}
\description{
`mixdown` with `stereo = FALSE`.
}
\examples{
render(s("bd*4"), 1)
}
