% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators.R
\name{rev}
\alias{rev}
\title{Reverse each cycle}
\usage{
rev(pat)
}
\arguments{
\item{pat}{A pattern.}
}
\value{
A pattern.
}
\description{
Mirror event times within every cycle. Masks `base::rev` when the package is attached.
}
\examples{
query_arc(rev(s("bd sd")), 0, 1)
}
