% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini.R
\name{s}
\alias{s}
\title{Mini-notation for drum sounds}
\usage{
s(text)
}
\arguments{
\item{text}{Mini-notation text.}
}
\value{
A pattern.
}
\description{
Parse text into a pattern of `list(sound = ...)` values, one cycle per input cycle.
}
\examples{
query_arc(s("bd sd"), 0, 1)
}
