% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{silence}
\alias{silence}
\title{A pattern with no haps}
\usage{
silence()
}
\value{
A pattern.
}
\description{
The empty pattern: queries always return an empty list.
}
\examples{
query_arc(silence(), 0, 1)
}
