% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators.R
\name{slow}
\alias{slow}
\title{Slow down a pattern}
\usage{
slow(pat, factor)
}
\arguments{
\item{pat}{A pattern.}

\item{factor}{Slow factor.}
}
\value{
A pattern.
}
\description{
Play the pattern `factor` times slower (fewer events per cycle).
}
\examples{
query_arc(slow(s("bd*4"), 2), 0, 1)
}
