% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators.R
\name{sometimes}
\alias{sometimes}
\title{Apply a transform on about half the cycles}
\usage{
sometimes(pat, fn)
}
\arguments{
\item{pat}{A pattern.}

\item{fn}{A function from pattern to pattern.}
}
\value{
A pattern.
}
\description{
`sometimes_by` with probability 0.5.
}
\examples{
query_arc(sometimes(s("bd*4"), function(p) rev(p)), 0, 2)
}
