% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators.R
\name{stack}
\alias{stack}
\title{Layer patterns}
\usage{
stack(...)
}
\arguments{
\item{...}{One or more patterns.}
}
\value{
A pattern.
}
\description{
Sound all input patterns at once. Masks `utils::stack` when the package is attached.
}
\examples{
query_arc(stack(s("bd*4"), s("~ sd")), 0, 1)
}
