% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{str_seed}
\alias{str_seed}
\title{Hash a string to an integer seed}
\usage{
str_seed(s)
}
\arguments{
\item{s}{A character string.}
}
\value{
An integer seed.
}
\description{
Simple string hash used to key deterministic randomness.
}
\examples{
str_seed("uzu:degrade:0")
}
