% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetry_gof.R
\name{symmetry_gof}
\alias{symmetry_gof}
\title{Goodness-of-fit test for symmetry.}
\usage{
symmetry_gof(ref_counts, var_counts, allelefreq)
}
\arguments{
\item{ref_counts}{Numeric list. Reference counts.}

\item{var_counts}{Numeric list. Variant counts.}

\item{allelefreq}{Number. Allele frequency.}
}
\value{
The p-value of chi-squared test for \code{ref_counts} and \code{var_counts}.
}
\description{
\code{symmetry_gof} calculates the chi-squared test for symmetry of allelic counts.
This tests whether the number of samples expressing more than 50% of the reference allele AND the amount of samples expressing less than 50% of the reference allele
correspond to the amounts we expect when assuming HWE AS WELL AS no allelic shift (on average equal expression of reference- and variant alleles in heterozygotes).
These are the expected number of homozygote ref-samples plus half the expected number of heterozygous samples,
and the expected number of homozygote var-samples plus half the expected number of heterozygous samples, both under HWE, respectively.
This function is used during imprinting analyses ONLY, as these are the assumptions made during those analyses.
}
\examples{
symmetry_gof(c(5, 8, 10, 3, 5, 6, 23, 4, 12, 10, 9, 6, 7, 25),
c(8, 8, 6, 4, 4, 10, 0, 7, 8, 4, 2, 7, 15, 13), 0.5)
symmetry_gof(c(5, 0, 0, 3, 5, 1, 23, 0, 12, 0, 9, 0, 1, 0),
c(1, 8, 6, 2, 0, 10, 0, 7, 0, 4, 0, 7, 0, 13), 0.5)
}
