% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators.R
\name{transpose}
\alias{transpose}
\title{Transpose}
\usage{
transpose(pat, semitones)
}
\arguments{
\item{pat}{A pattern.}

\item{semitones}{Semitones to shift, may be negative.}
}
\value{
A pattern.
}
\description{
Shift every note or `n` value by `semitones` semitones.
}
\examples{
note("c4") |> transpose(7)
}
