% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ts_duration}
\alias{ts_duration}
\title{Length of a TimeSpan}
\usage{
ts_duration(ts)
}
\arguments{
\item{ts}{A TimeSpan.}
}
\value{
A numeric duration.
}
\description{
Duration of a TimeSpan in cycles.
}
\examples{
ts_duration(timespan(0.25, 1))
}
