% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ts_intersection}
\alias{ts_intersection}
\title{Intersect two TimeSpans}
\usage{
ts_intersection(a, b)
}
\arguments{
\item{a}{A TimeSpan.}

\item{b}{A TimeSpan.}
}
\value{
A TimeSpan or NULL.
}
\description{
The overlap of two TimeSpans, or NULL when they do not overlap.
}
\examples{
ts_intersection(timespan(0, 1), timespan(0.5, 2))
}
