% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ts_split}
\alias{ts_split}
\title{Split a TimeSpan into equal parts}
\usage{
ts_split(ts, n)
}
\arguments{
\item{ts}{A TimeSpan.}

\item{n}{Number of parts.}
}
\value{
A list of n TimeSpans.
}
\description{
Divide a TimeSpan into n contiguous equal TimeSpans.
}
\examples{
ts_split(timespan(0, 1), 4)
}
