% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TumPur_LogLik_CPP4}
\alias{TumPur_LogLik_CPP4}
\title{DO SOMETHING}
\usage{
TumPur_LogLik_CPP4(
  q,
  a,
  b,
  ref_counts,
  var_counts,
  tumpur,
  weights,
  SCPthreshold,
  n = 0L,
  NumIntMethod = "Gregory",
  prec = 1e-04,
  Wvec = 0L,
  Nvec = 0L
)
}
\description{
DO SOMETHING
}
