% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{voice_bd}
\alias{voice_bd}
\title{Synthesized kick drum}
\usage{
voice_bd(dur, gain)
}
\arguments{
\item{dur}{Duration in seconds.}

\item{gain}{Gain multiplier.}
}
\value{
A numeric vector of samples.
}
\description{
Sine with a 150 to 50 Hz pitch drop and a fast decay.
}
\examples{
voice_bd(0.2, 0.8)
}
