% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{voice_cp}
\alias{voice_cp}
\title{Synthesized clap}
\usage{
voice_cp(dur, gain)
}
\arguments{
\item{dur}{Duration in seconds.}

\item{gain}{Gain multiplier.}
}
\value{
A numeric vector of samples.
}
\description{
Bandpassed noise with three bursts plus a tail.
}
\examples{
voice_cp(0.25, 0.8)
}
