% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{voice_rim}
\alias{voice_rim}
\title{Synthesized rim click}
\usage{
voice_rim(dur, gain)
}
\arguments{
\item{dur}{Duration in seconds.}

\item{gain}{Gain multiplier.}
}
\value{
A numeric vector of samples.
}
\description{
900 Hz square click plus highpassed noise.
}
\examples{
voice_rim(0.1, 0.8)
}
