% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{voice_sd}
\alias{voice_sd}
\title{Synthesized snare drum}
\usage{
voice_sd(dur, gain)
}
\arguments{
\item{dur}{Duration in seconds.}

\item{gain}{Gain multiplier.}
}
\value{
A numeric vector of samples.
}
\description{
180 Hz tone plus highpassed noise.
}
\examples{
voice_sd(0.2, 0.8)
}
