% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{voice_tom}
\alias{voice_tom}
\title{Synthesized tom}
\usage{
voice_tom(dur, gain)
}
\arguments{
\item{dur}{Duration in seconds.}

\item{gain}{Gain multiplier.}
}
\value{
A numeric vector of samples.
}
\description{
Sine with a 220 to 90 Hz pitch drop.
}
\examples{
voice_tom(0.25, 0.8)
}
