% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CppCnT_Optim}
\alias{CppCnT_Optim}
\title{DO SOMETHING}
\usage{
CppCnT_Optim(
  StartVals,
  ref_counts,
  var_counts,
  isCase,
  sprv,
  MemLim = 2048L,
  Xtra = 7L,
  step_size = 0.01,
  tol = 0.1,
  epsabs = 0.001
)
}
\description{
DO SOMETHING
}
