% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CppHetH0_Optim}
\alias{CppHetH0_Optim}
\title{Optimizer for theta for heterozygous samples assuming a fixed pi parameter in the beta-binomial mixture fit using expectation-maximization}
\usage{
CppHetH0_Optim(
  ThetaHetStart,
  probshift,
  ref_counts,
  var_counts,
  sprv,
  MemLim = 2048L,
  Xtra = 7L,
  step_size = 0.01,
  tol = 0.1,
  epsabs = 0.001
)
}
\description{
Optimizer for theta for heterozygous samples assuming a fixed pi parameter in the beta-binomial mixture fit using expectation-maximization
}
