% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMfit_betabinom_SEslack.R
\name{EMfit_betabinom_SEslack}
\alias{EMfit_betabinom_SEslack}
\title{This function is a work in progress. As such, it is not exported yet; but it allows for some leniency on the sequencing error parameter, hence "SESlack"}
\usage{
EMfit_betabinom_SEslack(
  data_counts,
  allelefreq = 0.5,
  SE,
  inbr = 0,
  dltaco = 10^-6,
  HWE = FALSE,
  p_InitEst = FALSE,
  ThetaInits = "moment",
  ReEstThetas = "moment",
  NoSplitHom = TRUE,
  NoSplitHet = TRUE,
  ResetThetaMin = 10^-10,
  ResetThetaMax = 10^-1,
  thetaTRY = c(10^-1, 10^-3, 10^-7),
  fitH0 = TRUE,
  SESlack = 10,
  probshift_InitEst = TRUE
)
}
\description{
This function is a work in progress. As such, it is not exported yet; but it allows for some leniency on the sequencing error parameter, hence "SESlack"
}
