% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_pi}
\alias{grad_pi}
\title{Beta-binomial gradient with respect to its pi parameter}
\usage{
grad_pi(ms, ns, pi, theta, MemLim = 2048L, Xtra = 7L)
}
\description{
Beta-binomial gradient with respect to its pi parameter
}
