% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maelstRom.R
\docType{package}
\name{maelstRom}
\alias{maelstRom-package}
\alias{maelstRom}
\title{maelstRom: A package for RNAseq-based allelic analyses}
\description{
Detection of imprinting, loss of imprinting, cis-eQTLS and differential allelic imbalance.
}
