% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MomentEst_MixedBetaBinom_SESlack.R
\name{MomentEst_MixedBetaBinom_SESlack}
\alias{MomentEst_MixedBetaBinom_SESlack}
\title{Similar to MomentEst_MixedBetaBinom, but allowing for slack on the SE parameter (see EMfit_betabinom_SEslack)}
\usage{
MomentEst_MixedBetaBinom_SESlack(
  ref_counts,
  var_counts,
  spr,
  spv,
  sprv,
  pi_RR_fix = NULL,
  pi_VV_fix = NULL,
  pi_het_fix = NULL
)
}
\description{
Similar to MomentEst_MixedBetaBinom, but allowing for slack on the SE parameter (see EMfit_betabinom_SEslack)
}
