% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TumPurHelpFun_CPP_R}
\alias{TumPurHelpFun_CPP_R}
\title{Internal helper function for TumPur_LogLik_CPP}
\usage{
TumPurHelpFun_CPP_R(
  RCcol,
  TC,
  TumReads_oi,
  a,
  b,
  q,
  SCP_oi,
  n,
  NumIntMethod,
  prec,
  Wvec,
  Nvec
)
}
\description{
Internal helper function for TumPur_LogLik_CPP
}
